+++ examples/resources/clumio_protection_group_bucket/basic.tf
@@ -1,0 +1,4 @@
+resource "clumio_protection_group_bucket" "basic_example" {
+  bucket_id   = "clumio-assigned-bucket-id"
+  protection_group_id = "protection-group-id"
+}